partitions_SUITE: Try to improve partial_false_positive
The testcase simulates a small delay between the times B sees A down
and C sees A down too. Because A is effectively down for both, B's
rabbit_node_monitor shouldn't see a partial partition.

When B sees A down first, it asks C about A. In the scenario the
testcase simulates, C tries to contact A, fails and reports to B that A
is down.

However, because of the way we simulate that delay, sometimes, C still
sees A when B asks about it, because the link between A and C isn't
blocked yet. When this happens, B decides there is a partial partition.
Therefore, the testcase fails because it tests that this shouldn't
happen.

To try to be closer to a real-world situation (where A would be really
down for B and C), we suspend C's rabbit_node_monitor until we block the
link between A and C. Thus B still asks C about A, but C only processes
the request when the link is indeed down.