partitions_SUITE: Try to improve partial_pause_minority
The testcase simulates a partial partition which should be detected by
the two nodes on each side of the partition. They should both decide
there is a partial partition and stop.

However, sometimes one of the node is faster than the other one: it sees
the partial partition and stops before the second node even sees the
partition. When this happens, the second node doesn't even bother to
stop (because the partition is full now). Thus the testcase fails
because it expects both nodes to stop.

This patch tries to sync partial partition checks so both A and B
decide there is a partial partition.